requirements/common.txt
Flask==3.0.3
google-cloud-discoveryengine
flet==0.23.2
hypercorn==0.17.3
google-cloud-firestore
plotly==5.23.0
google-cloud-aiplatform
google-auth[pyopenssl]
requests
